Move TestMain to a test file.
